api/server/httputils: compile with go < 1.7
